requirements.txt
anyio==3.6.2
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.6
crashtest==0.3.1
dependencies==7.7.0
distlib==0.3.6
dnspython==2.2.1
docker==6.0.1
dulwich==0.20.50
email-validator==1.3.0
fastapi==0.88.0
filelock==3.8.0
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.11.0
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==9.0.0
msgpack==1.0.4
orjson==3.8.2
packaging==21.3
pexpect==4.8.0
pkginfo==1.9.2
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycodestyle==2.10.0
pydantic==1.10.2
PyJWT==2.6.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
python-multipart==0.0.5
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.0
watchfiles==0.18.1
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
zipp==3.11.0
